Rearrange logging header stuff, add CMake flag
